Split IncomingRun out of MergeSchedule
Also update the credit docs for nominal credits.
